refactor room page
